Migrate file tool to maven
