+++ file-tool/src/main/java/rhit/Main.java
@@ -1,0 +1,13 @@
+package rhit;
+
+import java.io.IOException;
+
+public class Main {
+  public static void main(String[] args) throws IOException {
+    if (args.length > 0) {
+      new FileToolCli(args);
+    } else {
+      new FileToolGui();
+    }
+  }
+}
